requirements.txt
pymongo
python-dotenv
IMDbPY
cinemagoer
pandas
openai
requests
fastapi
uvicorn
pydantic
google-cloud-vision
torch 
torchvision 
pillow